requirements.txt
# local package
-e .

# external requirements
click~=8.1.7
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1

pandas~=1.5.3
PyYAML~=6.0.1
streamlit~=1.30.0
setuptools~=65.5.0